% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapis.R
\name{getAggregated}
\alias{getAggregated}
\title{Get aggregated data from LAPIS}
\usage{
getAggregated(session, ...)
}
\arguments{
\item{session}{The current session}

\item{...}{Additional filters}
}
\value{
The parsed content of the response
}
\description{
Makes a request to the /sample/aggregated endpoint of LAPIS and parses the response
}
\examples{
agg_data <- getAggregated(session, country = "Switzerland")
}
