% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapis.R
\name{getAminoAcidAlignment}
\alias{getAminoAcidAlignment}
\title{Get amino acid sequence alignment from LAPIS}
\usage{
getAminoAcidAlignment(
  session,
  gene,
  orderBy = NULL,
  limit = NULL,
  offset = NULL,
  downloadAsFile = T,
  compression = NULL,
  out = "aa_alignment",
  ...
)
}
\arguments{
\item{session}{The current session}

\item{gene}{The gene of interest. Valid values can be found in \code{session$genes}}

\item{orderBy}{The fields by which the sequences should be ordered. Available values: "gisaidEpiIsl", "random" and any gene in \code{session$genes}}

\item{limit}{Maximum number of sequences to align}

\item{offset}{Number of sequences to skip}

\item{downloadAsFile}{If the alignment should be written to file. If FALSE, the aligment is returned as string}

\item{compression}{Compression type. Available compression types: "gzip" and "zstd". Ignored if \code{downloadAsFile == FALSE}. If NULL, no compression}

\item{out}{Name of the output file (without file extension)}

\item{...}{Sequence filters.Valid filter keys can be found with \code{getFilters(session)}}
}
\value{
Amino acid sequence alignment in FASTA format
}
\description{
Makes a request to the /sample/alignedAminoAcidSequences/{gene} endpoint of LAPIS and parses the response
}
\examples{
getAminoAcidAlignment(session, "ORF1a", country = "Poland", dateDay=5, dateMonth=5, limit=3, out='aa_alignment')
}
