% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapis.R
\name{getDetails}
\alias{getDetails}
\title{Get metadata from LAPIS}
\usage{
getDetails(session, fields = NULL, limit = NULL, ...)
}
\arguments{
\item{session}{The current session}

\item{fields}{Fields to include}

\item{limit}{Maximal number of entries to return}

\item{...}{Additional filters}
}
\value{
Metadata table
}
\description{
Makes a request to the /sample/details endpoint of LAPIS and parses the response
}
\examples{
metadata <- getDetails(session, fields=c("country"), limit = 10, country = "Switzerland")
}
