% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapis.R
\name{getFilters}
\alias{getFilters}
\title{Get available sequence filters}
\usage{
getFilters(session)
}
\arguments{
\item{session}{The current session}
}
\value{
Array of available filter keys
}
\description{
Retrieves all filters available for a given session
}
\examples{
session <- initialize("https://lapis.cov-spectrum.org/gisaid/v2")
filters <- getFilters(session)
}
