% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapis.R
\name{getNucleotideMutations}
\alias{getNucleotideMutations}
\title{Get nucleotide mutations from LAPIS}
\usage{
getNucleotideMutations(session, ...)
}
\arguments{
\item{session}{The current session}

\item{...}{Filters}
}
\value{
Nucleotide mutations table
}
\description{
Makes a request to the /sample/nucleotideMutations endpoint of LAPIS and parses the response
}
\examples{
mutations <- getNucleotideMutations(session, region = "Europe", minProportion = 0.1)
}
