% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapis.R
\name{getNucleotideSequences}
\alias{getNucleotideSequences}
\title{Get nucleotide sequences from LAPIS}
\usage{
getNucleotideSequences(
  session,
  orderBy = NULL,
  limit = NULL,
  offset = NULL,
  downloadAsFile = TRUE,
  compression = NULL,
  out = "sequences",
  ...
)
}
\arguments{
\item{session}{The current session}

\item{orderBy}{The fields by which the sequences should be ordered. Available values: "main", "gisaidEpiIsl", "random"}

\item{limit}{Maximum number of sequences to include}

\item{offset}{Number of sequences to skip}

\item{downloadAsFile}{If the sequences should be written to file. If FALSE, the sequences are returned as string}

\item{compression}{Compression type. Available compression types: "gzip" and "zstd". Ignored if \code{downloadAsFile == FALSE}. If NULL, no compression}

\item{out}{Name of the output file (without file extension)}

\item{...}{Sequence filters. Valid filter keys can be found with \code{getFilters(session)}}
}
\value{
Nucleotide sequences in FASTA format
}
\description{
Makes a request to the /sample/unalignedNucleotideSequences endpoint of LAPIS and parses the response
}
\examples{
getNucleotideSequences(session, country = "Poland", dateDay = 5, dateMonth = 5, limit = 3, out = "lapis_sequences")
}
