% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapis.R
\name{getRequest}
\alias{getRequest}
\title{GET request to LAPIS}
\usage{
getRequest(session, endpoint, ...)
}
\arguments{
\item{session}{The current session}

\item{endpoint}{The endpoint where the request should be made}

\item{...}{Additional filters}
}
\value{
The parsed content of the response
}
\description{
Makes a GET request to a LAPIS endpoint and parses the response
}
\examples{
agg_data <- getRequest(session, "/sample/aggregated", country='Switzerland')
}
