% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapis.R
\name{initialize}
\alias{initialize}
\title{Initialize LAPIS session}
\usage{
initialize(host, expireOnUpdate)
}
\arguments{
\item{host}{URL of the host}

\item{expireOnUpdate}{If TRUE, the session expires as soon as the database receives an update}
}
\value{
A LAPIS session
}
\description{
Creates a session
}
\examples{
session <- initialize("https://lapis.cov-spectrum.org/gisaid/v2", expireOnUpdate = TRUE)
}
