% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapis.R
\name{runQuery}
\alias{runQuery}
\title{Querying LAPIS data}
\usage{
runQuery(session, endpoint, args)
}
\arguments{
\item{session}{The current session}

\item{endpoint}{The endpoint where the request should be made}

\item{args}{Arguments to be included in the request body}
}
\value{
The parsed content of the response
}
\description{
Makes a POST request to a LAPIS endpoint and parses the response
}
\examples{
data <- runQuery(session, "/sample/nucleotideMutations")
}
