% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapis.R
\name{getDetails}
\alias{getDetails}
\title{Get metadata from LAPIS}
\usage{
getDetails(
  session,
  fields = NULL,
  orderBy = NULL,
  limit = NULL,
  offset = NULL,
  ...
)
}
\arguments{
\item{session}{The current session}

\item{fields}{The fields to include in the response. Valid field names can be found in \code{session$metadata$name}. If NULL, all fields are returned.}

\item{orderBy}{The fields to order by; must be present in \code{fields}}

\item{limit}{Maximum number of sequences to include}

\item{offset}{Number of sequences to skip}

\item{...}{Filters. Valid filter keys can be found with \code{getFilters(session)}}
}
\value{
Sequence metadata
}
\description{
Makes a request to the /sample/details endpoint of LAPIS and parses the response
}
\examples{
metadata <- getDetails(session, fields = c("host", "age"), limit = 10, country = "Switzerland")
}
