% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapis.R
\name{getNucleotideInsertions}
\alias{getNucleotideInsertions}
\title{Get nucleotide insertions from LAPIS}
\usage{
getNucleotideInsertions(
  session,
  orderBy = NULL,
  limit = NULL,
  offset = NULL,
  ...
)
}
\arguments{
\item{session}{The current session}

\item{orderBy}{The fields to order by. Available values: "insertion", "count", "random"}

\item{limit}{Maximum number of insertions to include}

\item{offset}{Number of insertions to skip}

\item{...}{Sequence filters. Valid filter keys can be found with \code{getFilters(session)}}
}
\value{
Nucleotide insertions together with their counts among the sequences passing the filters
}
\description{
Makes a request to the /sample/nucleotideInsertions endpoint of LAPIS and parses the response
}
\examples{
insertions <- getNucleotideInsertions(session, region = "Europe", limit=10)
}
