% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapis.R
\name{getNucleotideMutations}
\alias{getNucleotideMutations}
\title{Get nucleotide mutations from LAPIS}
\usage{
getNucleotideMutations(
  session,
  orderBy = NULL,
  limit = NULL,
  offset = NULL,
  minProportion = 0.05,
  ...
)
}
\arguments{
\item{session}{The current session}

\item{orderBy}{The fields to order by. Available values: "mutation", "count", "proportion", "sequenceName", "mutationFrom", "mutationTo", "position", "random"}

\item{limit}{Maximum number of mutations to include}

\item{offset}{Number of mutations to skip}

\item{minProportion}{Minimal proportion required to include a mutation in the response}

\item{...}{Sequence filters. Valid filter keys can be found with \code{getFilters(session)}}
}
\value{
Nucleotide mutations together with their counts and proportions among the sequences passing the filters
}
\description{
Makes a request to the /sample/nucleotideMutations endpoint of LAPIS and parses the response
}
\examples{
mutations <- getNucleotideMutations(session, region = "Europe", minProportion = 0.1, limit = 10)
}
