% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapis.R
\name{getRequest}
\alias{getRequest}
\title{GET request to LAPIS}
\usage{
getRequest(url, args = NULL)
}
\arguments{
\item{url}{Request URL}

\item{args}{Additional filters}
}
\value{
The parsed content of the response
}
\description{
Makes a GET request to a LAPIS endpoint and parses the response
}
\examples{
agg_data <- getRequest("https://lapis.cov-spectrum.org/gisaid/v2/sample/aggregated", args=list(country='Switzerland'))
}
